Added a prompt to check if we want to delete lazy-lock.json. Most of the time it'll be no so defaulting to that when pressing enter.

+++ nuke-nvim.sh
@@ -1,4 +1,12 @@
-rm -f ~/.config/nvim/lazy-lock.json
+#!/usr/bin/env bash
+set -e
+
+read -rp "Also delete lazy-lock.json? [y/N] " answer
+if [[ "$answer" =~ ^[Yy]$ ]]; then
+  rm -f ~/.config/nvim/lazy-lock.json
+  echo "Deleted lazy-lock.json"
+fi
+
 rm -rf ~/.local/share/nvim
 rm -rf ~/.local/state/nvim
 rm -rf ~/.cache/nvim
